Add cucumber testing with a single feature
This commit sets up cucumber testing and brings in a single
"feature" initially "Create 'Contact' object". This is
probably the most complex test we have so will allow us to
drive out the maximum amount of complexity in a single
"outside -> in" process. By this we mean running the highest
level test and then filling in lower level tests and
implementation.